requirements.txt
tweepy
newspaper3k
beautifulsoup4
geopy
pymongo
python-dotenv
